feat(ux/security): optimize mobile UI responsiveness and secure file downloads
- Refactor student & BK dashboard tables to stacked card-style views on mobile (under md breakpoint) to prevent horizontal overflow.
- Remove Excel/Word export format choices from BK student export views, forcing PDF format via hidden input.
- Secure letter and archive attachments in production by replacing direct public storage asset paths with secure download controller routes featuring teacher authorization checks.
- Remove automatic user creation/hashing (Bcrypt) in student imports, manual creation, and edits to prevent PHP timeout errors.
- Add automatic class name normalization to consistent format (e.g. 7 andalan) via model mutator, migration, and feature tests.
- Disable database query logging and increase max execution time to 180s for student import routes.

+++ app/Http/Controllers/GuruBK/DashboardController.php
@@ -78,7 +78,7 @@ class DashboardController extends Controller
 
     public function show(Report $report)
     {
-        if ($report->handled_by !== Auth::id()) {
+        if ((int)$report->handled_by !== (int)Auth::id()) {
             return redirect()->route('gurubk.dashboard')->with('error', 'Anda tidak memiliki akses ke kasus ini.');
         }
 
@@ -119,7 +119,7 @@ class DashboardController extends Controller
 
     public function resolve(Report $report)
     {
-        if ($report->handled_by !== Auth::id()) {
+        if ((int)$report->handled_by !== (int)Auth::id()) {
             return back()->with('error', 'Anda tidak memiliki akses.');
         }
